multi_agent_weather/pyproject.toml
[project]
name = "multi-agent-weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.123.10",
    "google-adk[a2a]>=1.21.0",
    "uvicorn>=0.40.0",
]
